+++ ui-v2/app/components/consul-intention-form/index.hbs
@@ -1,12 +1,12 @@
-<form>
+<form onsubmit={{action 'submit' _item}}>
   <fieldset>
     <div role="group">
       <fieldset>
         <h2>Source</h2>
-        <label data-test-source-element class="type-text{{if item.error.SourceName ' has-error'}}">
+        <label data-test-source-element class="type-text{{if _item.error.SourceName ' has-error'}}">
           <span>Source Service</span>
           <PowerSelectWithCreate
-              @options={{services}}
+              @options={{_services}}
               @searchField="Name"
               @selected={{SourceName}}
               @searchPlaceholder="Type service name"
@@ -23,10 +23,10 @@
           <em>Search for an existing service, write in a future one, or write in any Service URI.</em>
         </label>
   {{#if (env 'CONSUL_NSPACES_ENABLED')}}
-        <label data-test-source-nspace class="type-text{{if item.error.SourceNS ' has-error'}}">
+        <label data-test-source-nspace class="type-text{{if _item.error.SourceNS ' has-error'}}">
           <span>Source Namespace</span>
           <PowerSelectWithCreate
-              @options={{nspaces}}
+              @options={{_nspaces}}
               @searchField="Name"
               @selected={{SourceNS}}
               @searchPlaceholder="Type namespace name"
@@ -46,10 +46,10 @@
       </fieldset>
       <fieldset>
         <h2>Destination</h2>
-        <label data-test-destination-element class="type-text{{if item.error.DestinationName ' has-error'}}">
+        <label data-test-destination-element class="type-text{{if _item.error.DestinationName ' has-error'}}">
           <span>Destination Service</span>
           <PowerSelectWithCreate
-              @options={{services}}
+              @options={{_services}}
               @searchField="Name"
               @selected={{DestinationName}}
               @searchPlaceholder="Type service name"
@@ -66,10 +66,10 @@
           <em>Search for an existing service or write in a future one.</em>
         </label>
   {{#if (env 'CONSUL_NSPACES_ENABLED')}}
-        <label data-test-destination-nspace class="type-text{{if item.error.DestinationNS ' has-error'}}">
+        <label data-test-destination-nspace class="type-text{{if _item.error.DestinationNS ' has-error'}}">
           <span>Destination Namespace</span>
           <PowerSelectWithCreate
-              @options={{nspaces}}
+              @options={{_nspaces}}
               @searchField="Name"
               @selected={{DestinationNS}}
               @searchPlaceholder="Type namespace name"
@@ -88,30 +88,30 @@
 {{/if}}
       </fieldset>
     </div>
-    <div role="radiogroup" class={{if item.error.Action ' has-error'}}>
+    <div role="radiogroup" class={{if _item.error.Action ' has-error'}}>
       {{#each (array 'allow' 'deny') as |intent|}}
         <label>
           <span>{{ capitalize intent }}</span>
-          <input type="radio" name="Action" value="{{intent}}" checked={{if (eq item.Action intent) 'checked'}} onchange={{ action 'change' }}/>
+          <input type="radio" name="Action" value="{{intent}}" checked={{if (eq _item.Action intent) 'checked'}} onchange={{ action 'change' }}/>
         </label>
       {{/each}}
     </div>
-    <label class="type-text{{if item.error.Description ' has-error'}}">
+    <label class="type-text{{if _item.error.Description ' has-error'}}">
       <span>Description (Optional)</span>
-      <input type="text" name="Description" value="{{item.Description}}" placeholder="Description (Optional)" onchange={{action 'change'}} />
+      <input type="text" name="Description" value="{{_item.Description}}" placeholder="Description (Optional)" onchange={{action 'change'}} />
     </label>
   </fieldset>
   <div>
-{{#if create }}
-    <button type="submit" {{ action "create" item}} disabled={{if (or item.isPristine item.isInvalid) 'disabled'}}>Save</button>
+{{#if _item.isNew }}
+    <button type="submit" disabled={{if (or _item.isPristine _item.isInvalid) 'disabled'}}>Save</button>
 {{ else }}
-    <button type="submit" {{ action "update" item}} disabled={{if item.isInvalid 'disabled'}}>Save</button>
+    <button type="submit" disabled={{if _item.isInvalid 'disabled'}}>Save</button>
 {{/if}}
-    <button type="reset" {{ action "cancel" item}}>Cancel</button>
-{{# if (and item.ID (not-eq item.ID 'anonymous')) }}
+    <button type="reset" onclick={{action oncancel _item}}>Cancel</button>
+{{# if (and _item.ID (not-eq _item.ID 'anonymous')) }}
     <ConfirmationDialog @message="Are you sure you want to delete this Intention?">
       <BlockSlot @name="action" as |confirm|>
-        <button data-test-delete type="button" class="type-delete" {{action confirm 'delete' item parent}}>Delete</button>
+        <button data-test-delete type="button" class="type-delete" {{action confirm ondelete _item}}>Delete</button>
       </BlockSlot>
       <BlockSlot @name="dialog" as |execute cancel message|>
         <DeleteConfirmation @message={{message}} @execute={{execute}} @cancel={{cancel}} />
