+++ ui-v2/app/styles/components/consul-intention-list.scss
@@ -1,0 +1,9 @@
+@import './consul-intention-list/index';
+.consul-intention-list {
+  @extend %consul-intention-list;
+}
+@media #{$--lt-wide-table} {
+  %consul-intention-list tr > :nth-last-child(2) {
+    display: none;
+  }
+}
