+++ ui-v2/app/styles/components/consul-intention-list/index.scss
@@ -1,0 +1,5 @@
+@import './skin';
+@import './layout';
+%consul-intention-list td.destination {
+  @extend %tbody-th;
+}
