+++ ui-v2/app/styles/components/index.scss
@@ -25,6 +25,7 @@
 @import './notice';
 @import './sort-control';
 @import './discovery-chain';
+@import './consul-intention-list';
 
 @import './tabular-details';
 @import './tabular-collection';
