+++ ui-v2/app/styles/components/table.scss
@@ -82,9 +82,6 @@ th span em {
   tr > .actions {
     display: none;
   }
-  html.template-intention.template-list tr > :nth-last-child(2) {
-    display: none;
-  }
   html.template-service.template-list tr > :last-child {
     display: none;
   }
